RF: block lists to check and test resolvers
- Move block list logic to checks
- Create IP resolver interfaces
  - Inject resolvers
- Add code cov file for code that just calls stdlib
