+++ checks/checks.go
@@ -4,10 +4,12 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/xray-web/web-check-api/checks/clients/ip"
 	"github.com/xray-web/web-check-api/checks/store/legacyrank"
 )
 
 type Checks struct {
+	BlockList  *BlockList
 	Carbon     *Carbon
 	IpAddress  *Ip
 	LegacyRank *LegacyRank
@@ -21,6 +23,7 @@ func NewChecks() *Checks {
 		Timeout: 5 * time.Second,
 	}
 	return &Checks{
+		BlockList:  NewBlockList(&ip.NetDNSLookup{}),
 		Carbon:     NewCarbon(client),
 		IpAddress:  NewIp(NewNetIp()),
 		LegacyRank: NewLegacyRank(legacyrank.NewInMemoryStore()),
