+++ handlers/rank_test.go
@@ -19,7 +19,7 @@ func TestHandleGetRank(t *testing.T) {
 		req := httptest.NewRequest(http.MethodGet, "/rank", nil)
 		rec := httptest.NewRecorder()
 
-		HandleBlockLists().ServeHTTP(rec, req)
+		HandleGetRank(nil).ServeHTTP(rec, req)
 
 		assert.Equal(t, http.StatusBadRequest, rec.Code)
 		assert.JSONEq(t, `{"error": "missing URL parameter"}`, rec.Body.String())
