+++ server/server.go
@@ -29,7 +29,7 @@ func (s *Server) routes() {
 
 	s.mux.Handle("GET /health", HealthCheck())
 
-	s.mux.Handle("GET /api/block-lists", handlers.HandleBlockLists())
+	s.mux.Handle("GET /api/block-lists", handlers.HandleBlockLists(s.checks.BlockList))
 	s.mux.Handle("GET /api/carbon", handlers.HandleCarbon(s.checks.Carbon))
 	s.mux.Handle("GET /api/cookies", handlers.HandleCookies())
 	s.mux.Handle("GET /api/dns-server", handlers.HandleDNSServer())
